requirements.txt
bandit==1.7.9
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
coverage==7.6.1
flake8==7.1.1
google-api-core==2.24.1
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
myql==1.2.7
oauthlib==3.2.2
packaging==24.2
pbr==6.1.1
pluggy==1.5.0
proto-plus==1.26.0
protobuf==5.29.3
pyaml==24.12.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.1
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
rauth==0.7.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
setuptools==75.8.2
six==1.17.0
stevedore==5.4.1
stringcase==1.2.0
uritemplate==4.1.1
urllib3==2.3.0
yahoo-oauth==2.1.0
yfpy==16.0.3
